Changed the reboot crontab to every 15 days, staggered.
The reason is that all the nodes were set to reboot at exactly the same time. This won’t work as it might create a window where no seed nodes are running when needed. The whole point of the seed nodes is high availability. At least a few will always be on when one is rebooting.