Added txindex to the lynx.conf
After working on the block explorers, I realized this flag needs to be set to get some additional data the block explorers need to run. Let’s assume this will always been needed since it’s a seed node. The first time the lynxd starts after this script is run is the equivalent of a -reindex, so no need to excplicitly specify it.