Removed all init variable questions, etc.
Abstracted the script by removing all requirements to run it as a StackScript at Linode. While it still can be (and is) called by a StackScript, it can also be used now as a init script on a home computer more easily. Default assumptions are made for the user. Default username and password for ssh is ‘lynx’ and ‘lynx’/