Gotta allow SSH access.
So I didn’t thik it was a good idea for a user to SSH into the RPI remotely via SSH while the build is happeneding cause if it gets interrupted, it might not recover. So, the ssh file isn’t touched till the end. The user COULD log in via KVM, but as long as they don’t start messing with the build, it should be okay. WIth the limited RAM, it’s best not to be messing with it till it’s done.

Signed-off-by: Ben <ben@getlynx.io>