Locked Pi user account.
We are no longer using the Pi user account. On any version of LynxCI now, either on the Pi or in the cloud, the default login is lynx | lynx for the username and password. Also, we have a friendly message after login that tells if the node is accessible from the outside.