Updated messages and flagged wallet OFF in config
This way, a LynxCI node can be easily turned on to enable wallet functions, but by default the wallet functions are off. For now, this is a more secure setting. Especially since the RPC port is public and open and the only thing preventing access is the RPC credentials. This is far more safe.